requirements.txt
tensorflow==2.10
matplotlib==3.8.4
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
